Move product category resolvers into dedicated modules
