+++ back/boxtribute_server/graph_ql/bindables.py
@@ -6,6 +6,8 @@ from ariadne import InterfaceType, ObjectType, UnionType
 from ..beneficiary.fields import beneficiary
 from ..beneficiary.mutations import mutation as beneficiary_mutation
 from ..beneficiary.queries import query as beneficiary_query
+from ..core.product_category.fields import product_category
+from ..core.product_category.queries import query as product_category_query
 from ..core.size_range.fields import size_range
 from ..metrics.fields import metrics
 from ..metrics.queries import query as metrics_query
@@ -14,13 +16,13 @@ from ..user.fields import user
 from ..user.queries import query as user_query
 
 # Container for QueryTypes
-query_types = (beneficiary_query, metrics_query, user_query)
+query_types = (beneficiary_query, metrics_query, product_category_query, user_query)
 
 # Container for MutationTypes
 mutation_types = (beneficiary_mutation,)
 
 # Container for ObjectTypes (public as immutable tuple)
-_object_types = [beneficiary, metrics, size_range, user]
+_object_types = [beneficiary, metrics, product_category, size_range, user]
 
 
 def _register_object_type(name):
@@ -41,7 +43,6 @@ classic_location = _register_object_type("ClassicLocation")
 organisation = _register_object_type("Organisation")
 packing_list_entry = _register_object_type("PackingListEntry")
 product = _register_object_type("Product")
-product_category = _register_object_type("ProductCategory")
 qr_code = _register_object_type("QrCode")
 shipment = _register_object_type("Shipment")
 shipment_detail = _register_object_type("ShipmentDetail")
